Factor out is_path to own file
