+++ include/boost/nowide/fstream.hpp
@@ -9,9 +9,11 @@
 #define BOOST_NOWIDE_FSTREAM_HPP_INCLUDED
 
 #include <boost/nowide/config.hpp>
+#include <boost/nowide/detail/is_path.hpp>
 #include <boost/nowide/filebuf.hpp>
 #include <istream>
 #include <ostream>
+#include <utility>
 
 namespace boost {
 namespace nowide {
@@ -60,8 +62,6 @@ namespace nowide {
                  int FileBufType = BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT>
         class fstream_impl;
 
-        template<typename Path, typename Result>
-        struct enable_if_path;
     } // namespace detail
     /// \endcond
 
@@ -94,9 +94,8 @@ namespace nowide {
         }
 
         template<typename Path>
-        explicit basic_ifstream(
-          const Path& file_name,
-          typename detail::enable_if_path<Path, std::ios_base::openmode>::type mode = std::ios_base::in)
+        explicit basic_ifstream(const Path& file_name,
+                                detail::enable_if_path_t<Path, std::ios_base::openmode> mode = std::ios_base::in)
         {
             open(file_name, mode);
         }
@@ -143,9 +142,8 @@ namespace nowide {
             open(file_name, mode);
         }
         template<typename Path>
-        explicit basic_ofstream(
-          const Path& file_name,
-          typename detail::enable_if_path<Path, std::ios_base::openmode>::type mode = std::ios_base::out)
+        explicit basic_ofstream(const Path& file_name,
+                                detail::enable_if_path_t<Path, std::ios_base::openmode> mode = std::ios_base::out)
         {
             open(file_name, mode);
         }
@@ -200,8 +198,8 @@ namespace nowide {
         }
         template<typename Path>
         explicit basic_fstream(const Path& file_name,
-                               typename detail::enable_if_path<Path, std::ios_base::openmode>::type mode =
-                                 std::ios_base::in | std::ios_base::out)
+                               detail::enable_if_path_t<Path, std::ios_base::openmode> mode = std::ios_base::in
+                                                                                              | std::ios_base::out)
         {
             open(file_name, mode);
         }
@@ -314,8 +312,8 @@ namespace nowide {
                 open(file_name.c_str(), mode);
             }
             template<typename Path>
-            typename detail::enable_if_path<Path, void>::type open(const Path& file_name,
-                                                                   std::ios_base::openmode mode = T_StreamType::mode())
+            detail::enable_if_path_t<Path, void> open(const Path& file_name,
+                                                      std::ios_base::openmode mode = T_StreamType::mode())
             {
                 open(file_name.c_str(), mode);
             }
@@ -357,41 +355,6 @@ namespace nowide {
 #ifdef BOOST_MSVC
 #pragma warning(pop)
 #endif
-        /// Trait to heuristically check for a *\::filesystem::path
-        /// Done by checking for make_preferred and filename member functions with correct signature
-        template<typename T>
-        struct is_path
-        {
-            using one = char;
-            struct two
-            {
-                char dummy[2];
-            };
-
-            template<typename U, U& (U::*)(), U (U::*)() const>
-            struct Check;
-            template<typename U>
-            static one test(Check<U, &U::make_preferred, &U::filename>*);
-            template<typename U>
-            static two test(...);
-
-            enum
-            {
-                value = sizeof(test<T>(0)) == sizeof(one)
-            };
-        };
-        template<bool B, typename T>
-        struct enable_if
-        {};
-        template<typename T>
-        struct enable_if<true, T>
-        {
-            using type = T;
-        };
-        /// SFINAE trait which has a member "type = Result" if the Path is a *\::filesystem::path
-        template<typename Path, typename Result>
-        struct enable_if_path : enable_if<is_path<Path>::value, Result>
-        {};
     } // namespace detail
 } // namespace nowide
 } // namespace boost
